Don't escape HTML for the organiser info.
This fixes up the display of organiser information on the website. This
is POTENTIALLY an XSS hole, as we're presuming the full_name,
twitter_url, and mobile number are HTML-safe. However, this is only for
organisers, and organisers are a) trusted users and b) set by an admin.
So in practice, this should be fine.