Better tests for the InvitationManager and don't email if we don't create a SessionInvitation.
This started with Despo spotting a potential issue that could result in
users getting spammed:
https://github.com/codebar/planner/pull/179#discussion_r22188398

This seemed like a reasonable thing to test, so I wrote a couple of
tests - one to ensure that it would email, and one to ensure that it
wouldn't. Surprisingly, the former test also failed, which led me to
discover that most of the invitation_manager_spec wasn't really testing
anything.

The core of the problem was that the fabricated :students group had no
members - which meant that the assertions set in the .each blocks
weren't being run, so these test cases were essentially empty. Adding
members to the fabricated groups revealed some bugs in these test cases;
these have now been fixed, along with the initial potential-spam bug.