Fix access control on member controller; redirect users to new wizard flow.
The logged_in? method doesn't work as a before_action filter, but
authenticate_member! will. This changeset uses that instead. It also
redirects users to the new wizard on login.

We check whether we're on a page that lets the user update their details
before redirecting, so we don't end up in an infinite redirect.