Schedule an AllocateSpacesJob when the allocation time changes
The jobs will sort themselves out if they find they are running too
early or too late, and they are idempotent so it doesn't matter if
they run multiple times. We just need to stick them on the queue every
time a change happens, and the right one will run.