Implements workshop deletion feature
A workshop is eligible for deletion only if there aren’t any
invitations sent and if it is created within ‘X’ hours from the current
time. As of this commit, the value of X is assumed to be 4 hours(based
on the discussion here -
https://github.com/codebar/planner/issues/371#issuecomment-169356782).
The timecop gem is added as it makes testing wrt the current time more
easier. The changes in this commit address the issue -
https://github.com/codebar/planner/issues/371