Fix link to chapter on profile page
The member profile page lists all the current users' subscriptions. But
the entries link to the admin section of the site, which most people
won't have access to.

People clicking the link will see a message "You can't be here"
(indicating that they haven't got enough permissions for this).

I think the original intention was to link to the chapter page. This
commit fixes the link.