Review changes as per @dansteele feedback
Moved the `helper_method` calls above the private method to make it
clearer that they're not private. They're tricky to unit test from the
controller spec (they're called in the view). As the feature spec
touches them, I clear the controller spec.

Moved the method that finds all the members in a chapter and sorts by
attendance to the Chapter model. I need to find a better way of
approaching it, but the current DB structure confuses the heck out of
me. Either way it has a spec that'll seed the DB so I can work on it.