Fix intermittently failing test
The order or the `attendees` isn't guaranteed in
`WorkshopPresenter#attendees_emails`, which means this test sometimes
fails, since `expect(a).to eq(b)` expects the array to be exactly the
same. Unlike `match_array`, which only verifies that the arrays have
identical items.

For example, this test failed in the unrelated dependabot upgrade for
`pundit`:

https://travis-ci.org/codebar/planner/builds/440397050
https://github.com/codebar/planner/pull/827