Create button disappears when new sponsor form has error #907
When saving a new sponsor fails, the 'new' form is rendered.
The submit button text is dependent on the params action either being new or edit. When rendered, the params action is now create, so no submit button is shown.

Added a test to manage_sponsor_spec that confirms that there is a 'Create sponsor' button upon initial submission with invalid input.
Updated sponsors/_form to show the 'Create sponsor' button if the action is new OR create.