Remove PhantomJS/Poltergeist as currently unused
There are no feature tests which currently use `js: true`. The only one
that needs it (in `spec/features/admin/manage_workshop_attendances_spec.rb`)
is currently skipped, and also doesn't work with the current
Poltergeist/PhantomJS setup anyway.

PhantomJS is currently unmaintained:
https://github.com/ariya/phantomjs/issues/15105#issuecomment-322850178.
It looks like the next thing to do is to look into using headless
Chrome, which will be addressed in a future PR.