Add support for JS feature tests w/Chrome headless
This commit leans very heavily on a blogpost from GitLab:
https://about.gitlab.com/2017/12/19/moving-to-headless-chrome/

This commit doesn't add some of the additional tweaks as found in
https://gitlab.com/gitlab-org/gitlab-ce/blob/master/spec/support/capybara.rb
such as timeouts and access to browser logs - these can be added as
needed in future.

Tests can be run with a visible (headful?) browser by passing
CHROME_HEADLESS=false (or "no" or "0") to a test: e.g.
`CHROME_HEADLESS=false bundle exec rspec`

This will probably not work with TravisCI yet, this will be addressed in
a future commit (along with additional documentation)