Fix: flaky test job scopes #published returns all published job
  - Error seen as:
     - Job scopes #published returns all published job
       Failure/Error: expect(Job.published.all).to eq(approved)

  - Problem line:
    expect(Job.published.all).to eq(approved)

  - Why?
    - Making bad assumptions about DB ordering
    - Job.published.all is something like this in SQL:
      select * from jobs where status ==  2 # aka published
        - select *often* returns records by order but it makes no
          guarantee and frequently not in order
      However, rspec eq asserts by contents and order
      [1, 2, 3] eq [1, 2, 3] => returns true
      [1, 2, 3] eq [1, 3, 2] => returns false

  - Conclusion
    the spec will fail because SQL can supply jobs in any order
    while the rspec eq expects it in an exact order

------------------------------------------------------------
Why have they been deleted?

Test wise all we needed to know was that the Class contained
an enum called status and importantly, the order was draft,
pending and published. Important because the database only
holds numbers and if someone was to switch the enum order
the database wouldn't know of the switch and everything would
break. This is now covered with the shoulda method -
define_enum_for