(feature) restrict access to admin/jobs and admin/feedback to members with the admin role
As the admin/application_controller checks if a user is either an admin or an organiser for now add a super_admin/application_controller.

This could benefit from additional refactoring so there is better separation of what the organiser and admin roles can do. The tests should ideally also be refactored to reflect this.