(enhancement) add `day_temporal_pronoun` to event presenter
This should be able to correctly return today or tomorrow so that email reminders correctly indicate when an event is taking place.

As the rake task is executed between 30-4 hours before the workshop takes place, the only two choices are these so we don’t care about any other scenarios.