Ignore invites_sent column from now on
Before we remove it, this commit will help us to stop considering
it whenever we’re about to send invites and will also stop updating
it. After talking with Despo, what we have in place at the moment
seems enough to take care of avoiding duplicated invites.

`MeetingInvite`s seems to have a unique constraint in place that
will prevent us from running into duplicated invites created and
sent, so this commit adds a test to take that into consideration
when admins try to send meeting invites to users that have already
been invited