Display error message when invitation update fails
Alex documented on #1474 the issue of adding someone manually from
the waitlist, sometimes the member gets the correct email but they
disappear from the list they were supposed to be added to.

Despo pointed me to this particular case there the invitation update
could fail and we would have no way to know it. When this happens,
their spot on the waiting list would be deleted.

This commit addresses that issue now, paying attention to the result
of the update, notifying the admin if it failed and only removing
the WaitList entity if no error occurred.