Improve HTML re-rendering after marked as attended
This change will slightly improve the time we have to wait when
checking someone in for a workshop. In order to get faster times,
the whole technique has to be replaced by something different.
At the moment, there are a lot of JS events and layout reflows
which slow down the page considerably.