Don't allow rsvp-ing a workshop that is in the past #1484
The backend did not do validation on the workshop's time and date when
a user sent an rsvp. There is validation in the view when the workshop
page is rendered, so the button to attend does not show up on past
workshops. However if the page was old and the button was there, a user
could still sign up and get an invitation created. The workshop was then
displayed as in the past.

For virtual workshops it is possible to attend anyway, so a confirmation
email would be sent. This change stops the system from ever creating
invitations for past workshops.