Solves a weird issue with Capybara and testing
I was getting some 404 issues on a single Capybara test, and to be
honest I did what I could to understand it. Since I could not
really solve it and it was blocking an important fix, I decide to
tell Capybara not to break our tests because *test* generated
images were not available.