Show today’s workshop(s) as upcoming
As an admin, something that is constantly annoying to me is losing
quick access to the workshop I’m organising today because it is
past 6.30pm and it no longer shows up in the “Upcoming workshops”
section.

Adding a new scope that includes today’s workshops as well, so that
organisers have access to the workshop at hands while running it.