Add last_overridden_id property to schema
This will allow us to record the ID of the last organiser that
overrode a WorkshopInvite. I gave it a thought about this and
tried a few different approaches, but this seemed the best compromise.

I’ve looked into keeping track of every change through a few
gems, but this would only grow the DB without much advantage,
I believe. I don’t think we need (at the moment, at least) a
detailed record of all the changes.

At its core, I’m just interested in knowing if a WorkshopInvitation
has been “tampered with”. This should help us with that.