Skip students and coaches without TOC accepted
After talking with Kim about this a few weeks ago, we realised it
doesn’t make much sense to email members that have not accepted
our terms about new events. We see that as not having come back to
codebar in a while, thus making more sense to save our resources.