Make default avatar a square
I've used a blank square from Wikipedia as a default for the avatar. This is a little brittle, because it assumes wikipedia will always be around and that we can use the square. We could alter this to be a square that we own, if we wanted. This should be changed as soon as we've properly fixed the logo upload