Explicitly create the Omniauth hash
I notice that the test fails in Github with the error

```
undefined method `[]' for nil:NilClass
     # ./app/controllers/auth_services_controller.rb:48:in `create'
     # ./spec/features/member_portal_spec.rb:95:in `block (3 levels) in <top (required)>'
     # ./spec/spec_helper.rb:114:in `block (2 levels) in <top (required)>'
```

I think this is because we don't set the Gihub auth mock. It could be set elsewhere, and that might be why the test is flakey. I've explicitly set it here to try to get the tests working consistently