Use local path to Apple Touch icons
This uses the <link> tag, which can tell Apple Safari about the icons that it requests to know about.

In addition to changing the elements in <head>, this also adds copies of Icon-128.png at Apple's well-known file locations.

These are two ways of telling Apple Safari about that icon.