Fix: ensure chrome driver not used for non js tests
My previous implementation set the default driver after each test but this was not resolving the issue because chrome was being set as the default driver after js:true tests ran when rack_test should have been used instead.  This commit resolves that by setting the current_driver in spec_helper instead so that before a js:true test the current_driver is set to the javascript_driver = chrome and after every test the current_driver is set to the default-driver (rack_test). This should ensure that the chrome driver isnt retained incorrectly after each js test.