Fix failing tests due to error in new query
I thought the issue was related to the data fabrication but actually the issue was in the query I made.  It was not filtering by workshop id.  The query returned an array of workshops and the first workshop was picked and so when two workshops were fabricated the sponsors were always the same.  I've resolved this by filtering by workshop id in the query on the workshop model