Have MemberPresenter return false to undefined methods
In order to use the MemberPresenter more flexibly, I want to remove the constant checking. The Presenter should be able to intialise with a `nil` object and then deal with the consequences of that. This approach will return `false` to any method not defined on the presenter or the presented object.

Signed-off-by: jonathan.kerr <3410350+jonodrew@users.noreply.github.com>