replace change_table with add_column in migration
I was running into a `TypeError: can't quote Array (TypeError)`
error when running this migration and at first I thought using
`change_table` instead of `add_column` seemed to help.

In actuality the error still occurs when this migration is run as part
of a sequence of multiple migrations, but the migration runs
successfully when run on its own.

However, the issue seems to be fixed in Rails 7.1 so:

- can switch this to use `add_column` which is clearer in terms of the
  intent, and
- wait until our Rails 7.1 upgrade is merged before shipping this