fix(logging): implement semantic logger
we currently run out of log quota on a daily basis, so this
change aims to condense the log lines so we don't. :)

therefor we now use the rails_semantic_logger gem as its purpose
is to help with that. this commit also replaces rails_12factor
(which is why it's removed) since it would otherwise conflict with
the new gem.

logging config for all "environments" gets appeneded in application.rb
now.

resolves: #2339